+++ src/community/people/components/molecules/UserDetailsCentered/UserDetailsCentered.tsx
@@ -4,13 +4,13 @@ import { useDropzone } from "react-dropzone";
 
 import { useGetUploadedImage } from "~community/common/api/FileHandleApi";
 import Avatar from "~community/common/components/molecules/Avatar/Avatar";
+import { appModes } from "~community/common/constants/configs";
 import { accountPageTestId } from "~community/common/constants/testIds";
 import { FileTypes } from "~community/common/enums/CommonEnums";
 import { theme } from "~community/common/theme/theme";
 import { usePeopleStore } from "~community/people/store/store";
 import { ModifiedFileType } from "~community/people/types/AddNewResourceTypes";
 import { EmployeeDetails } from "~community/people/types/EmployeeTypes";
-import { ProfileModes } from "~enterprise/common/enums/CommonEum";
 import { useGetEnvironment } from "~enterprise/common/hooks/useGetEnvironment";
 
 interface Props {
@@ -72,7 +72,7 @@ const UserDetailsCentered: FC<Props> = ({
     FileTypes.USER_IMAGE,
     cardData?.authPic,
     true,
-    environment !== ProfileModes.ENTERPRICE
+    environment !== appModes.ENTERPRISE
   );
 
   const handleUnSelectPhoto = (): void => {
