+++ src/fallback/auth/enterpriseAuthOptions.ts
@@ -1,0 +1,5 @@
+import { NextAuthOptions } from "next-auth";
+
+export const enterpriseAuthOptions: NextAuthOptions = {
+  providers: []
+};
