+++ src/fallback/common/api/authApi.ts
@@ -1,0 +1,8 @@
+export const useCheckLoginMethod = (
+  onSuccess: (response: any) => void,
+  onError?: (error: unknown) => void
+) => {
+  return {
+    mutate: async (data: any) => {}
+  };
+};
