+++ src/fallback/common/api/setDeviceTokenApi.ts
@@ -1,0 +1,1 @@
+export const setDeviceToken = (fcmToken: string): void => {};
