+++ src/fallback/common/assets/languages/english/english.ts
@@ -1,0 +1,1 @@
+export const english = {};
