+++ src/fallback/common/configs/firebase.ts
@@ -1,0 +1,1 @@
+export const firebaseConfig = {};
