+++ src/fallback/common/constants/dbKeys.ts
@@ -1,0 +1,2 @@
+export const isProdMaintenanceMode = "";
+export const isNonProdMaintenanceMode = "";
