+++ src/fallback/common/hooks/useFCMToken.ts
@@ -1,0 +1,9 @@
+const useFcmToken = () => {
+  const token = null;
+
+  const resetUnreadCount = () => {};
+
+  return { token, resetUnreadCount };
+};
+
+export default useFcmToken;
