+++ src/fallback/common/hooks/useGetEnvironment.ts
@@ -1,0 +1,4 @@
+export const useGetEnvironment = () => {
+  const env = process.env.NEXT_PUBLIC_MODE;
+  return env;
+};
