+++ src/fallback/common/hooks/useS3Download.ts
@@ -1,0 +1,8 @@
+const useS3Download = () => {
+  const s3FileUrls: Record<string, string> = {};
+  const downloadS3File = (key: string) => {};
+
+  return { s3FileUrls, downloadS3File };
+};
+
+export default useS3Download;
