+++ src/fallback/common/utils/monitoring.ts
@@ -1,0 +1,1 @@
+export const initializeHotjar = () => {};
