+++ src/fallback/people/api/CheckUserLimitApi.ts
@@ -1,0 +1,6 @@
+export const useCheckUserLimit = (isEnterpriseMode: boolean) => {
+  return {
+    data: false,
+    isSuccess: true
+  };
+};
