+++ src/fallback/people/api/GlobalLoginMethodApi.ts
@@ -1,0 +1,8 @@
+export const useGetGlobalLoginMethod = (
+  isEnterpriseMode: boolean,
+  tenantName: string
+) => {
+  return {
+    data: ""
+  };
+};
