+++ src/fallback/people/api/peopleApi.ts
@@ -1,0 +1,10 @@
+import { EmployeeRoleLimit } from "../types/EmployeeTypes";
+
+export const useGetEmployeeRoleLimit = (
+  onSuccess: (response: EmployeeRoleLimit) => void,
+  onError: (error: unknown) => void
+) => {
+  return {
+    mutate: async () => {}
+  };
+};
