+++ src/fallback/people/types/EmployeeTypes.ts
@@ -1,0 +1,9 @@
+export interface EmployeeRoleLimit {
+  leaveAdminLimitExceeded: boolean;
+  attendanceAdminLimitExceeded: boolean;
+  peopleAdminLimitExceeded: boolean;
+  leaveManagerLimitExceeded: boolean;
+  attendanceManagerLimitExceeded: boolean;
+  peopleManagerLimitExceeded: boolean;
+  superAdminLimitExceeded: boolean;
+}
